proxy/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.5
httpx>=0.27.1
pydantic-settings>=2.5.2
asyncpg==0.29.0
boto3>=1.35.0
python-multipart>=0.0.9
mcp[cli]>=1.9.0
PyJWT[crypto]>=2.8.0
mcpauth>=0.1.0
